simplifying session service - work in progress
