+++ test-review/solutions/microservice-session/src/main/java/com/redhat/training/conference/speaker/SpeakerFromService.java
@@ -1,6 +1,4 @@
-package org.acme.conference.session;
-
-import javax.json.bind.annotation.JsonbCreator;
+package com.redhat.training.conference.speaker;
 
 public class SpeakerFromService {
     String uuid;
@@ -13,7 +11,6 @@ public class SpeakerFromService {
         this.nameLast = nameLast;
     }
 
-    @JsonbCreator
     public static SpeakerFromService of(String uuid, String nameFirst, String nameLast) {
         return new SpeakerFromService(uuid, nameFirst, nameLast);
     }
