+++ test-review/solutions/microservice-session/src/main/java/com/redhat/training/conference/speaker/SpeakerRepository.java
@@ -1,9 +1,9 @@
-package org.acme.conference.session;
+package com.redhat.training.conference.speaker;
 
 import javax.enterprise.context.ApplicationScoped;
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
 
 @ApplicationScoped
 public class SpeakerRepository implements PanacheRepository<Speaker> {
-  
+
 }
