--- test-review/solutions/microservice-session/src/main/java/org/acme/conference/session/LivenessCheck.java
@@ -1,16 +1,0 @@
-package org.acme.conference.session;
-
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.eclipse.microprofile.health.Liveness;
-
-import javax.enterprise.context.ApplicationScoped;
-
-@Liveness
-@ApplicationScoped
-public class LivenessCheck implements HealthCheck {
-    @Override
-    public HealthCheckResponse call() {
-        return HealthCheckResponse.up("Service is alive");
-    }
-}
