--- test-review/solutions/microservice-session/src/main/java/org/acme/conference/session/ReadinessCheck.java
@@ -1,16 +1,0 @@
-package org.acme.conference.session;
-
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.eclipse.microprofile.health.Readiness;
-
-import javax.enterprise.context.ApplicationScoped;
-
-@Readiness
-@ApplicationScoped
-public class ReadinessCheck implements HealthCheck {
-    @Override
-    public HealthCheckResponse call() {
-        return HealthCheckResponse.up("Service is ready");
-    }
-}
