--- test-review/solutions/microservice-session/src/main/java/org/acme/conference/session/SessionRepository.java
@@ -1,9 +1,0 @@
-package org.acme.conference.session;
-
-import javax.enterprise.context.ApplicationScoped;
-import io.quarkus.hibernate.orm.panache.PanacheRepository;
-
-@ApplicationScoped
-public class SessionRepository implements PanacheRepository<Session> {
-  
-}
