+++ test-review/solutions/microservice-session/src/test/java/com/redhat/training/conference/session/SessionFakeFactory.java
@@ -1,4 +1,4 @@
-package org.acme.conference.session;
+package com.redhat.training.conference.session;
 
 public class SessionFakeFactory {
